mcsdk-dylan-config: add latest config for use with meta-mcsdk layer
The plan for MCSDK is to use meta-mcsdk layer on top of meta-ti and meta-arago
for staging early development/integration and last-minute fixes in the form of
bbappends, instead of keeping forked copies of meta-ti and meta-arago with own
changes. Ultimately, all those changes should be upstreamed into mainline
meta-ti and meta-arago, making meta-mcsdk almost empty. This flow is meant to
simplify development and maintenance of MCSDK and allow automatic inheriting
of Yocto Project compatibility status from mainline meta-ti and meta-arago.

Signed-off-by: Denys Dmytriyenko <denys@ti.com>