Add git_retry.sh wrapper around git
We are seeing a sporadic failure in our nightly builds
due to the git clone/fetch calls timing out and erroring.
These events are usually resolved by just trying the git
call again, but in a scripted flow it causes a failure
in that run.  The git_retry.sh script solves that by
allowing the git command to be retried a few times while
sleeping in between.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>