configs: Switch github repos from git to https
This is a follow on commit to the previous one that changes all
of the older configs as well so that they remain usable without
users needing to do anything.

github announced removal of support of unencrypted git:// last
year[1], However, we still use git:// in many of our configurations.
Switch them over to https://

[1]
https://github.blog/2021-09-01-improving-git-protocol-security-github/i

Signed-off-by: Ryan Eatmon <reatmon@ti.com>