oe-layersetup: Add config option for inclusive vars
Part of the transition to kirkstone was to rename some variables
to align with community standards.  Since oe-layersetup needs to
work with both dunfell and kirkstone, we need a way to control
which branch we are pointing to so that we can put the correct
variables in the conf/setenv file.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>