oe-layertool-setup.sh: Remove all control over number of threads
It was pointed out that bitbake now picks good defaults for the
BB_NUMBER_THREADS and PARALLEL_MAKE variables, so we should remove
all control over those variables from the script level.  Also
comment out the default settings in the sample conf files so that
bitbake can be in control, but if the user needs to override them
they have the example of what to change.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>