sample-files: gitconfig.sample: Stop using git:// urls
We did this in meta-ti and meta-arago but didn't propagate back into
oe-layersetup yet so I'll give it a go. git:// is no longer functional
proxies over on latest git versions (example: poky crops container for
ubuntu 22.04[1]), also git:// is highly susceptible to man-in-the-middle
attacks[2]. Dump git:// in favor of https.

[1] https://hub.docker.com/r/crops/poky/tags 22.04
[2] https://github.blog/2021-09-01-improving-git-protocol-security-github/

Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>