configs: Stop using git:// urls
We did this in meta-ti and meta-arago but I didn't propagate back into
oe-layersetup yet so I'll give it a go. git:// is no longer functional
proxies over on latest git versions (example: poky crops container for
ubuntu 22.04), also git:// is highly susceptible to man-in-the-middle
attacks[1]. Dump git:// in favor of https.

NOTE: We leave the current release configurations as they are, and
we depend on the recommended git-config example to do the re-routing

[1] https://github.blog/2021-09-01-improving-git-protocol-security-github/

Signed-off-by: Randolph Sapp <rs@ti.com>
[nm@ti.com: Updates to commit message and subject to indicate improvements]
Signed-off-by: Nishanth Menon <nm@ti.com>
Reviewed-by: Denys Dmytriyenko <denys@konsulko.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>