git_retry.sh: Sleep progressively longer in between tries
We are seeing some odd failures with fetching bitbake.  Most of the
builds in a given run are fine, but a couple will fail to fetch bitbake.
We are attempting to debug what is causing it, but one thing to try is
sleeping for longer periods of time in between git failures.  If it is
the remote server that is having issues, then give it longer and longer
to recover.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>