src/templates/local*: Clean-up whitespace at end of the line
A whitespace at the end of a line can cause patches sent to the mailing
list to drop those when applying the patch, but make will keep trying
to add it back in.  Need to watch for this future.  Maybe even add
checks during make to fail if we see it.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>