configs/arago-scarthgap-*-config: qt5 -> qt6
Move the rest of the scarthgap configs to use qt6 instead of qt5 not
that scarthgap-wip and scarthgap-next have been tested. Use the 9.6
branch for now builds since this is the latest supported version tested
against scarthgap.

Moving forward TI builds of arago will enable and test Qt6 by default,
but unlike before this will not be done in a way that creates an
explicit dependency. This should allow customers and layers that consume
or extend arago to toggle Qt support more easily.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>