git_retry.sh: remove let, use arithmetic evaluation
Since none of this script is checking the exit code of let, we can
replace it with a standard arithmetic evaluation and switch the shebang
to a POSIX shell.

According to bash(1) let simply performs an arithmetic evaluation with
an additional check that returns 1 if the last arg evaluates to 0.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>