kas: fetch bitbake as part of the setup
We can get kas to fetch it's own version of bitbake with a little
coercion. Adjust the template to populate the appropriate repo section
for bitbake.

This behaves like any other repo definition, with the key difference
being it has no enabled layers. This means you must explicitly set the
root layer "." to disabled.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>