kas: do not add a default target if one is not given
There are configs, such as the distroless configs, that may not specify
a default target. In those cases we should not have a "target" key in
our kas config. An unspecified target is fine, but an empty target is a
violation of their schema.

Detected through the use of yamllint's trailing whitespace check.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>