src: Rework bitbake-setup support
Recent changes to the upcoming bitbake-setup tool require a tweak to the
json files and layer files.  This commit updates the src files needed to
handle the change.

The big change since the last pass is that bitbake-setup no longer
requires a layer with templates for each config to list out the layers
and local.conf files.  The current json supports the layers and config
fragment files in the oe-layersetup layer for the local.conf settings.

- Add more detail to the distro xml files and rename them to reflect the
additional data in the file.
- Add the list of machines that meta-ti-bsp and meta-beagle support.
- Update the convert script to reflect the file name change above.
- Update the Makefile to stop generating the templates/ files.
- Remove the xslt for the templates/ files.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>